refactor(pr9): move model PGM out of package to docs/; drop daft runtime dep
Move build_nested_tcri_pgm / draw_tcri_pgm_nested out of tcri.utils into
docs/model_pgm.py (daft is now a docs-only tool, not a tcri runtime dependency —
removed from pyproject). import tcri no longer imports daft. Suite green.
